+++ src/log/log.js
@@ -156,6 +156,7 @@ class VConsoleLogTab extends VConsolePlugin {
     window.console.warn = this.console.warn;
     window.console.debug = this.console.debug;
     window.console.error = this.console.error;
+    window.console.clear = this.console.clear;
     this.console = {};
   }
 
@@ -237,6 +238,7 @@ class VConsoleLogTab extends VConsolePlugin {
       methodList.map(function(method) {
         that.console[method] = window.console[method];
       });
+      that.console.clear = window.console.clear;
     }
 
     methodList.map(method => {
@@ -247,6 +249,11 @@ class VConsoleLogTab extends VConsolePlugin {
         });
       };
     });
+
+    window.console.clear = (...args) => {
+      that.clearLog();
+      that.console.clear.apply(window.console, args);
+    };
   }
 
   clearLog() {